2007-04-04  Kyle Galloway  <kgallowa@redhat.com>
	* classpath/gnu/classpath/jdwp/util/VariableTable.java: Change longs
	to ints for argCnt and slots.
	(write): Replace writeLong with writeInt for the above.



git-svn-id: svn+ssh://gcc.gnu.org/svn/gcc/trunk@123511 138bc75d-0d04-0410-961f-82ee72b054a4

+++ libjava/gnu/classpath/jdwp/util/VariableTable.h
@@ -33,8 +33,8 @@ public:
   VariableTable(jint, jint, JArray< jlong > *, JArray< ::java::lang::String * > *, JArray< ::java::lang::String * > *, JArray< jint > *, JArray< jint > *);
   virtual void write(::java::io::DataOutputStream *);
 private:
-  jlong __attribute__((aligned(__alignof__( ::java::lang::Object)))) argCnt;
-  jlong slots;
+  jint __attribute__((aligned(__alignof__( ::java::lang::Object)))) argCnt;
+  jint slots;
   JArray< jlong > * lineCI;
   JArray< jint > * slot;
   JArray< jint > * lengths;
